maj DAO produit + ajout produitController et fichiers jsp

+++ 20_projet_spring_ecommerce/src/fr/adaming/model/Categorie.java
@@ -66,7 +66,7 @@ public class Categorie {
 	}
 
 	/* _____________________ Getters & Setters ________________________ */
-	public long getIdCategorie() {
+	public int getIdCategorie() {
 		return idCategorie;
 	}
